+++ ClickHouse/S01-install-and-init.R
@@ -4,21 +4,29 @@ library(TKCat)
 ## Configuration ----
 
 TKCAT_HOME <- "/data/pgodard/Projects/TKCat_UCB_TBN"
+# TKCAT_HOME <- "/data/pgodard/Projects/TKCat_Test"
 TKCAT_NAT_PORT <- 9101
+# TKCAT_NAT_PORT <- 9301
 TKCAT_HTTP_PORT <- 9111
+# TKCAT_HTTP_PORT <- 9311
 
-SRC_DIR <- here::here("ClickHouse", "ClickHouse-Files")
+USERFILE <- system.file(
+   "ClickHouse/users.xml",
+   package="TKCat"
+)
 CH_VERSION <- "21.4.6"
 
 CONTAINER <- "ucb_tbn_tkcat"
-INSTANCE_NAME <- "UCB - TBN"
+# CONTAINER <- "test_tkcat"
+# INSTANCE_NAME <- "UCB - TBN"
+INSTANCE_NAME <- "Test"
 ADMIN <- "pgodard"
 CONTACT <- "Patrice Godard <patrice.godard@ucb.com>"
 
 ###############################################################################@
 ## Check ----
-if(!file.exists(SRC_DIR)){
-   stop("The ClickHouse-Files directory does not exist")
+if(!file.exists(USERFILE)){
+   stop(sprintf("%s does not exist", USERFILE))
 }
 if(file.exists(TKCAT_HOME)){
    stop(sprintf("%s already exists", TKCAT_HOME))
@@ -34,11 +42,8 @@ Sys.chmod(TKCAT_DATA, mode="777", use_umask=FALSE)
 TKCAT_CONF <- file.path(TKCAT_HOME, "conf")
 dir.create(TKCAT_CONF, recursive=TRUE)
 Sys.chmod(TKCAT_CONF, mode="777", use_umask=FALSE)
-TKCAT_CONFIG <- file.path(TKCAT_CONF, "config.xml")
 TKCAT_USERS <- file.path(TKCAT_CONF, "users.xml")
-file.copy(file.path(SRC_DIR, "config.xml"), TKCAT_CONFIG)
-Sys.chmod(TKCAT_CONFIG, mode="777", use_umask=FALSE)
-file.copy(file.path(SRC_DIR, "users-init.xml"), TKCAT_USERS)
+file.copy(USERFILE, TKCAT_USERS)
 Sys.chmod(TKCAT_USERS, mode="777", use_umask=FALSE)
 
 ###############################################################################@
@@ -50,7 +55,6 @@ system(paste(
    sprintf('--volume %s:/var/lib/clickhouse', TKCAT_DATA),
    sprintf('--publish=%s:9000', TKCAT_NAT_PORT),
    sprintf('--publish=%s:8123', TKCAT_HTTP_PORT),
-   sprintf('--volume %s:/etc/clickhouse-server/config.xml', TKCAT_CONFIG),
    sprintf('--volume %s:/etc/clickhouse-server/users.xml', TKCAT_USERS),
    '--restart=always',
    sprintf('yandex/clickhouse-server:%s', CH_VERSION)